% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{CancerInSilico}
\alias{CancerInSilico}
\alias{CancerInSilico-package}
\title{CancerInSilico}
\description{
\tabular{ll}{
Package: \tab CellModel\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2016-06-24\cr
License: \tab LGPL\cr
}
}
\author{
Maintainer: Elana J. Fertig \email{ejfertig@jhmi.edu},
            Thomas D. Sherman \email{tsherma4@jhu.edu}
}

