% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R
\docType{class}
\name{CellModel-class}
\alias{CellModel-class}
\title{CellModel}
\description{
The top-level CellModel class. All other cell model
 classes inherit from this in some way
}
\section{Slots}{

\describe{
\item{\code{cells}}{A list object that describes the state of the cells
at each time. The state representation depends on the type of
model run, and is accessed by the function designed for each
model type.}

\item{\code{initialNum}}{number of cells at time 0}

\item{\code{runTime}}{number of model hours to run the simulation}

\item{\code{density}}{initial density of cells}

\item{\code{boundary}}{keep cells within circular boundary}

\item{\code{syncCycles}}{start all cells in the beginning of interphase}

\item{\code{randSeed}}{random seed used for both R and C++ functions}

\item{\code{outputIncrement}}{how often simulation info is displayed}

\item{\code{recordIncrement}}{how often cell info is recorded (controls size
of resulting CellModel object}

\item{\code{timeIncrement}}{controls how fine the model timestep is}

\item{\code{cellTypes}}{list of CellType objects used in the model}

\item{\code{cellTypeInitFreq}}{initial frequency of cell types among cells}

\item{\code{drugs}}{list of Drug objects used in the model}
}}

