% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{checkDataSet}
\alias{checkDataSet}
\title{verify gene expression data set is valid for this package}
\usage{
checkDataSet(dataSet, genes)
}
\arguments{
\item{dataSet}{matrix of gene expression data where row names are genes}

\item{genes}{names of all genes being simulated}
}
\value{
no value is return, but errors/warnings are thrown related to
 potential problems in the data set
}
\description{
checks a data set before it is used to calibrate the 
 pathway values for min/max expression
}

