% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{combineGeneExpression}
\alias{combineGeneExpression}
\title{Combine Gene Expression Matrices}
\usage{
combineGeneExpression(expList, combineFUN = max)
}
\arguments{
\item{expList}{list of expression matrices}

\item{combineFUN}{function to use when combining multiple values}
}
\value{
matrix containing combined expression values
}
\description{
combines mutliple matrices by in a similiar way to rbind,
but combines muttiple values for a single gene by a given function
}
\keyword{internal}

