% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMeanExpression.R
\name{combinePathways}
\alias{combinePathways}
\title{\code{combinePathways} Combine gene expression values for all simulated pathways}
\usage{
combinePathways(pathSimOutput, combineFUN = max, ...)
}
\arguments{
\item{pathSimOutput}{List of matrices containing gene by time expression values for each pathway generated from \code{simulateMeanExpression}.}

\item{combineFUN}{Optional; Function used to combine expression values from genes shared in multiple pathways. Defaults to maximum value in any condition.}

\item{...}{other input arguments to combineFUN}
}
\value{
Matrix simulating the mean gene expression values for pathway genes from a \code{\link{CellModel}} simulation at the specified frequency in \code{samplingFrequency}.
}
\description{
\code{combinePathways} Combine gene expression values for all simulated pathways
}

