% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DrasdoHohmeModel.R
\docType{class}
\name{DrasdoHohmeModel-class}
\alias{DrasdoHohmeModel-class}
\title{DrasdoHohmeModel}
\description{
Implementation of an off-latice cell-based model
based on the work in Drasdo, Hohme (2003)
}
\section{Slots}{

\describe{
\item{\code{nG}}{number of monte carlo steps between each growth trial}

\item{\code{epsilon}}{constant that controls the probability trails are accepted}

\item{\code{delta}}{controls distance over which short range interactions occur}
}}

