% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Drug.R
\docType{class}
\name{Drug-class}
\alias{Drug-class}
\title{Drug}
\description{
describes the properties of a drug
}
\section{Slots}{

\describe{
\item{\code{name}}{name of drug}

\item{\code{timeAdded}}{the time at which this drug is added to the simulation}

\item{\code{cycleLengthEffect}}{effect this drug has on cell cycle length}
}}

