% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GeneExpressionParams.R
\docType{class}
\name{GeneExpressionParams-class}
\alias{GeneExpressionParams-class}
\title{GeneExpressionParams}
\description{
Parameters for simulating gene expression
}
\section{Slots}{

\describe{
\item{\code{sampleFreq}}{how often to generate data}

\item{\code{RNAseq}}{generate RNA-seq data}

\item{\code{singleCell}}{generate single cell data}

\item{\code{nCells}}{number of cells to sample at each time point}

\item{\code{nDummyGenes}}{number of dummy genes}

\item{\code{dummyDist}}{function to determine expression of dummy genes}

\item{\code{combineFUN}}{function used to combine gene expression data}

\item{\code{randSeed}}{random seed}

\item{\code{perError}}{error for normal error model}

\item{\code{bcvCommon}}{error for voom error model}

\item{\code{bcvDF}}{degrees of freedom for voom error model}

\item{\code{dropoutPresent}}{whether to simulate dropout in single cell data}

\item{\code{dropoutMid}}{parameter for dropout distribution}

\item{\code{dropoutShape}}{parameter for dropout distribution}
}}

