% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OffLatticeModel.R
\docType{methods}
\name{getAxisAngle}
\alias{getAxisAngle}
\alias{getAxisAngle,OffLatticeModel-method}
\title{get cell axis angle at a given time}
\usage{
getAxisAngle(model, time, cell)

\S4method{getAxisAngle}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
axis angle
}
\description{
get cell axis angle at a given time
}
\examples{
data(SampleModels)
getAxisAngle(modDefault, modDefault@runTime, 1)
}

