% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getAxisAngle}
\alias{getAxisAngle}
\title{\code{getAxisAngle} get the axis angle of each cell}
\usage{
getAxisAngle(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
vector containing the axis angle of each cell at time
}
\description{
\code{getAxisAngle} get the axis angle of each cell
}

