% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OffLatticeModel.R
\docType{methods}
\name{getAxisLength}
\alias{getAxisLength}
\alias{getAxisLength,OffLatticeModel-method}
\title{get cell axis length at a given time}
\usage{
getAxisLength(model, time, cell)

\S4method{getAxisLength}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
axis length
}
\description{
get cell axis length at a given time
}
\examples{
data(SampleModels)
getAxisLength(modDefault, modDefault@runTime, 1)
}
