% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getAxisLength}
\alias{getAxisLength}
\title{\code{getAxisLength} get the axis length of each cell}
\usage{
getAxisLength(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
vector containing the axis length of each cell at time
}
\description{
\code{getAxisLength} get the axis length of each cell
}

