% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getCellDistance}
\alias{getCellDistance}
\alias{getCellDistance,OffLatticeModel-method}
\title{get distance between two cells}
\usage{
getCellDistance(model, time, cellA, cellB)

\S4method{getCellDistance}{OffLatticeModel}(model, time, cellA, cellB)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cellA}{id of cell to query}

\item{cellB}{id of cell to query}
}
\value{
distance between cellA and cellB
}
\description{
get distance between two cells
}
\examples{
data(SampleModels)
getCellDistance(modDefault, modDefault@runTime, 1, 2)
}

