% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getCellPhase}
\alias{getCellPhase}
\alias{getCellPhase,OffLatticeModel-method}
\title{get phase of a cell at a given time}
\usage{
getCellPhase(model, time, cell)

\S4method{getCellPhase}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
cell phase
}
\description{
get phase of a cell at a given time
}
\examples{
data(SampleModels)
getCellPhase(modDefault, modDefault@runTime, 1)
}

