% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellPhasePos.R
\name{getCellPhasePos}
\alias{getCellPhasePos}
\title{\code{getCellPhasePos} Determines which cells are in the process of dividing}
\usage{
getCellPhasePos(model, time, opt = 1)
}
\arguments{
\item{model}{A CellModel}

\item{time}{Time}

\item{opt}{changes between getting mitosis or synthesis phase}
}
\value{
the size of the cell population over time
}
\description{
\code{getCellPhasePos} Determines which cells are in the process of dividing
}

