% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getCellType}
\alias{getCellType}
\alias{getCellType,OffLatticeModel-method}
\title{get type of a cell at a given time}
\usage{
getCellType(model, time, cell)

\S4method{getCellType}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
cell type
}
\description{
get type of a cell at a given time
}
\examples{
data(SampleModels)
getCellType(modDefault, modDefault@runTime, 1)
}

