% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getCellTypes}
\alias{getCellTypes}
\title{\code{getCellTypes} get a list of all the cell types in the model}
\usage{
getCellTypes(model, time)
}
\arguments{
\item{model}{a CellModel object}
}
\value{
a list of factors
}
\description{
\code{getCellTypes} get a list of all the cell types in the model
}
\examples{
getCellTypes(runCancerSim(1,1), 20)
}

