% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OffLatticeModel.R
\docType{methods}
\name{getCoordinates}
\alias{getCoordinates}
\alias{getCoordinates,OffLatticeModel-method}
\title{get coordinates of a cell at a given time}
\usage{
getCoordinates(model, time, cell)

\S4method{getCoordinates}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
pair of (x,y) coordinates
}
\description{
get coordinates of a cell at a given time
}
\examples{
data(SampleModels)
getCoordinates(modDefault, modDefault@runTime, 1)
}
