% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getCoordinates}
\alias{getCoordinates}
\title{\code{getCoordinates} get a two dimensional matrix of all the cell coordinates}
\usage{
getCoordinates(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
an N X 2 matrix of cell coordinates at time
}
\description{
\code{getCoordinates} get a two dimensional matrix of all the cell coordinates
}

