% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getCycleLength}
\alias{getCycleLength}
\alias{getCycleLength,OffLatticeModel-method}
\title{get cycle length of a cell at a given time}
\usage{
getCycleLength(model, time, cell)

\S4method{getCycleLength}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
cycle length in hours
}
\description{
get cycle length of a cell at a given time
}
\examples{
data(SampleModels)
getCycleLength(modDefault, modDefault@runTime, 1)
}
