% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCycleLengthDistribution.R
\name{getCycleLengthDistribution}
\alias{getCycleLengthDistribution}
\title{\code{getCycleLengthDistribution} Plots a distribution of the growth
     rates of all the cells}
\usage{
getCycleLengthDistribution(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{The time of interest for the growth rates}
}
\value{
returns a vector of growth rates for cells alive at time
}
\description{
\code{getCycleLengthDistribution} Plots a distribution of the growth
     rates of all the cells
}
\examples{
getCycleLengthDistribution(runModel(1,10),0)
}

