% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getCycleLengths}
\alias{getCycleLengths}
\title{\code{getCycleLengths} return the cycle lengths of each cells at time}
\usage{
getCycleLengths(model, time)
}
\arguments{
\item{model}{a CellModel object}

\item{time}{time of interest}
}
\value{
the cycle lengths of each cell at time
}
\description{
\code{getCycleLengths} return the cycle lengths of each cells at time
}

