% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensity.R
\docType{methods}
\name{getDensity,CellMatrix-method}
\alias{getDensity,CellMatrix-method}
\title{\code{getDensity} Gets the density of cells at a given time}
\usage{
\S4method{getDensity}{CellMatrix}(mat, time)
}
\arguments{
\item{mat}{A Cell Model object}

\item{time}{The time of interest}
}
\value{
The density of cells at that time
}
\description{
\code{getDensity} Gets the density of cells at a given time
}

