% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDensity.R
\docType{methods}
\name{getDensity,CellModel-method}
\alias{getDensity,CellModel-method}
\title{\code{getDensity} Gets the density of cells at a given time}
\usage{
\S4method{getDensity}{CellModel}(model, time)
}
\arguments{
\item{model}{A Cell Model}

\item{time}{The time of interest}
}
\value{
The density of cells at that time (not quite the same as confluency)
}
\description{
\code{getDensity} Gets the density of cells at a given time
}
\examples{
getDensity(runCancerSim(1,10),10)
}

