% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getDensity}
\alias{getDensity}
\alias{getDensity,OffLatticeModel-method}
\title{get density of the cell population at a given time}
\usage{
getDensity(model, time)

\S4method{getDensity}{OffLatticeModel}(model, time)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}
}
\value{
density
}
\description{
get density of the cell population at a given time
}
\examples{
data(SampleModels)
getDensity(modDefault, modDefault@runTime)
}
