% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getDensity}
\alias{getDensity}
\title{\code{getDensity} gets the density of cells at a given time}
\usage{
getDensity(model, time)
}
\arguments{
\item{model}{A Cell Model}

\item{time}{time in model hours}
}
\value{
The density of cells at that time (not quite the same as confluency)
}
\description{
\code{getDensity} gets the density of cells at a given time
}
\examples{
getDensity(runCancerSim(1,1),1)
}

