% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDividingCells.R
\name{getDividingCells}
\alias{getDividingCells}
\title{\code{getDividingCells} Determines which cells are in the process of dividing}
\usage{
getDividingCells(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{Time}
}
\value{
the size of the cell population over time
}
\description{
\code{getDividingCells} Determines which cells are in the process of dividing
}

