% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugEffect.R
\name{getDrugEffect}
\alias{getDrugEffect}
\title{\code{getDrugEffect}}
\usage{
getDrugEffect(slowDown, cycleLengthSeq)
}
\arguments{
\item{cycleLengthSeq}{A sequence spanning the range of cycle lengths}

\item{slwoDown}{A vector specifying the slow down of the growth rate}
}
\value{
A list of vectors specifying the distribution of drug effects depending on the growth rate of the cell
}
\description{
\code{getDrugEffect}
}
\examples{
getDrugEffect(0.3, seq(2,12,0.1))
}

