% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{getGROWTHexpression}
\alias{getGROWTHexpression}
\title{\code{getGROWTHexpression} calculate gene expression for a pathway effected by the growth rate of the cells}
\usage{
getGROWTHexpression(model, cells, time)
}
\arguments{
\item{model}{A CellModel}

\item{cells}{the indices of cells to calculate expression for}

\item{time}{model time in hours}
}
\value{
gene expression for each cell
}
\description{
\code{getGROWTHexpression} calculate gene expression for a pathway effected by the growth rate of the cells
}
