% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrowthRateDistribution.R
\docType{methods}
\name{getGrowthRateDistribution,CellModel-method}
\alias{getGrowthRateDistribution,CellModel-method}
\title{\code{getGrowthRateDistribution} Plots a distribution of the growth
     rates of all the cells}
\usage{
\S4method{getGrowthRateDistribution}{CellModel}(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{The time of interest for the growth rates}
}
\value{
returns a vector of growth rates for cells alive at time
}
\description{
\code{getGrowthRateDistribution} Plots a distribution of the growth
     rates of all the cells
}
\examples{
getGrowthRateDistribution(runModel(100,10),10)
}

