% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getGrowthRates}
\alias{getGrowthRates}
\title{\code{getGrowthRates} get the model growth rates of each cell}
\usage{
getGrowthRates(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
vector containing the growth rate of each cell at time
}
\description{
\code{getGrowthRates} get the model growth rates of each cell
}

