% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{getGtoMexpression}
\alias{getGtoMexpression}
\title{\code{getGtoMexpression} calculate gene expression for a pathway effected by the G to M transition}
\usage{
getGtoMexpression(model, cells, time, time_window)
}
\arguments{
\item{model}{A CellModel}

\item{cells}{the indices of cells to calculate expression for}

\item{time}{model time in hours}

\item{time_window}{the window of time (model hours) to check if a cell made the transition}
}
\value{
gene expression for each cell
}
\description{
\code{getGtoMexpression} calculate gene expression for a pathway effected by the G to M transition
}

