% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getLocalDensity}
\alias{getLocalDensity}
\alias{getLocalDensity,OffLatticeModel-method}
\title{get neighborhood density around a cell at a given time}
\usage{
getLocalDensity(model, time, cell, radius)

\S4method{getLocalDensity}{OffLatticeModel}(model, time, cell, radius)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}

\item{radius}{distance to search for neighboring cells}
}
\value{
density
}
\description{
get neighborhood density around a cell at a given time
}
\examples{
data(SampleModels)
getLocalDensity(modDefault, modDefault@runTime, 1, 3.3)
}

