% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{getMeanExp}
\alias{getMeanExp}
\title{Calculate Mean Expression Value}
\usage{
getMeanExp(pathways, activity)
}
\arguments{
\item{pathways}{list of pathway objects}

\item{activity}{list of pathway activity}
}
\value{
matrix of mean expression values
}
\description{
calculates the mean expression values for each gene and sample
by combining the pathway activity with the gene expression range
}
\keyword{internal}

