% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfCells.R
\docType{methods}
\name{getNumberOfCells,CellModel-method}
\alias{getNumberOfCells,CellModel-method}
\title{\code{getNumberOfCells} Plots the total number of cells vs time}
\usage{
\S4method{getNumberOfCells}{CellModel}(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{the time of interest}
}
\value{
the size of the cell population over time
}
\description{
\code{getNumberOfCells} Plots the total number of cells vs time
}
\examples{
getNumberOfCells(runCancerSim(1,10),10)
}

