% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getNumberOfCells}
\alias{getNumberOfCells}
\alias{getNumberOfCells,OffLatticeModel-method}
\title{get number of cells in the model at a given time}
\usage{
getNumberOfCells(model, time)

\S4method{getNumberOfCells}{OffLatticeModel}(model, time)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}
}
\value{
number of cells
}
\description{
get number of cells in the model at a given time
}
\examples{
data(SampleModels)
getNumberOfCells(modDefault, modDefault@runTime)
}
