% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNumberOfCells.R
\name{getNumberOfCells}
\alias{getNumberOfCells}
\title{\code{getNumberOfCells} return the total number of cells at time}
\usage{
getNumberOfCells(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{the time of interest}
}
\value{
the size of the cell population
}
\description{
\code{getNumberOfCells} return the total number of cells at time
}
\examples{
getNumberOfCells(runModel(1,10),10)
}

