% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getNumNeighbors}
\alias{getNumNeighbors}
\title{\code{getNumNeighbors} return the number of neighboring cells}
\usage{
getNumNeighbors(model, time, index)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}

\item{index}{index of cell}
}
\value{
the number of neighbors around the cell at index
}
\description{
\code{getNumNeighbors} return the number of neighboring cells
}
