% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwayExpressionValues.R
\name{getPathwayExpressionValues}
\alias{getPathwayExpressionValues}
\title{\code{getPathwayExpressionValues} Determine the gene expression values to use for each gene in each pathway.}
\usage{
getPathwayExpressionValues(ReferenceDataset = NULL, lambda = 1/3, pathways)
}
\arguments{
\item{lambda}{Optional; Parameter of the exponential distribution used to determine the maximum expression value of each simulated gene in the pathway. Defaults to 1/3. Not used if values are determined from a dataset in \code{ReferenceDataSet}.}

\item{pathway}{A list of genes associated with pathways.}

\item{ReferenceDataSet}{Optional; Reference gene expression dataset to use to calculate the gene expression values for genes in the pathway to use as mean values in the simulation. Defaults values randomly selected from an exponential distribution with parameter \code{lambda}. If specified by the user, \code{row.names} of the ReferenceDataset must match gene names in the \code{pathway} argument. In this case, gene expression values will be set as the values for genes in the pathway of the sample with max median value of raw RNA expression and expression Z-score for pathway genes.}
}
\value{
list of numeric objects for the maximum expression value of each gene in each pathway.
}
\description{
\code{getPathwayExpressionValues} Determine the gene expression values to use for each gene in each pathway.
}

