% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwaysSim.R
\name{getPathwaysSim}
\alias{getPathwaysSim}
\title{\code{getPathwaysSim} Determine the pathways from which to simulate gene expression data and check its accuracy}
\usage{
getPathwaysSim(pathways)
}
\arguments{
\item{pathway}{Optional; A list of genes associated with pathways, defaulting to gene symbols in the package data object \code{\link{inSilicoPathways}}. If specified by the user, names of pathways should be GtoM for genes associated with the G to M cell cycle transition, GtoS for genes associated with G to S, Prox for genes associated with contact, and Growth for genes associated with growth factor receptor signaling.}
}
\value{
A list of genes associated with pathways that will be used for the simulation.
}
\description{
\code{getPathwaysSim} Determine the pathways from which to simulate gene expression data and check its accuracy
}

