% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{getPROXexpression}
\alias{getPROXexpression}
\title{\code{getPROXexpression} calculate gene expression for a pathway effected by neighboring cells}
\usage{
getPROXexpression(model, cells, time)
}
\arguments{
\item{model}{A CellModel}

\item{cells}{the indices of cells to calculate expression for}

\item{time}{model time in hours}
}
\value{
gene expression for each cell
}
\description{
\code{getPROXexpression} calculate gene expression for a pathway effected by neighboring cells
}

