% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getRadii}
\alias{getRadii}
\title{\code{getRadii} get the radius of each cell}
\usage{
getRadii(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
vector containing the radius of each cell at time
}
\description{
\code{getRadii} get the radius of each cell
}

