% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OffLatticeModel.R
\docType{methods}
\name{getRadius}
\alias{getRadius}
\alias{getRadius,OffLatticeModel-method}
\title{get cell radius at a given time}
\usage{
getRadius(model, time, cell)

\S4method{getRadius}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
radius of cell
}
\description{
get cell radius at a given time
}
\examples{
data(SampleModels)
getRadius(modDefault, modDefault@runTime, 1)
}
