% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalCells.R
\docType{methods}
\name{getTotalCells,CellModel-method}
\alias{getTotalCells,CellModel-method}
\title{\code{getTotalCells} Plots the total number of cells vs time}
\usage{
\S4method{getTotalCells}{CellModel}(model)
}
\arguments{
\item{model}{A CellModel}
}
\value{
the size of the cell population over time
}
\description{
\code{getTotalCells} Plots the total number of cells vs time
}
\examples{
getTotalCells(runModel(10,100))
}

