% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTotalCells.R
\name{getTotalCells}
\alias{getTotalCells}
\title{\code{getTotalCells} Plots the total number of cells vs time}
\usage{
getTotalCells(model)
}
\arguments{
\item{model}{A CellModel}
}
\value{
the size of the cell population over time
}
\description{
\code{getTotalCells} Plots the total number of cells vs time
}
\examples{
getTotalCells(runModel(10,100))
}

