% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{getTrialAcceptRate}
\alias{getTrialAcceptRate}
\alias{getTrialAcceptRate,OffLatticeModel-method}
\title{get rate of growth acceptance of a given cell at a given time}
\usage{
getTrialAcceptRate(model, time, cell)

\S4method{getTrialAcceptRate}{OffLatticeModel}(model, time, cell)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}

\item{cell}{id of cell to query}
}
\value{
acceptance rate
}
\description{
get rate of growth acceptance of a given cell at a given time
}
\examples{
data(SampleModels)
getTrialAcceptRate(modDefault, modDefault@runTime, 1)
}

