% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R
\docType{methods}
\name{initialize,CellModel-method}
\alias{initialize,CellModel-method}
\title{Constructor for CellModel}
\usage{
\S4method{initialize}{CellModel}(.Object, initialNum, runTime, density,
  boundary = 1, syncCycles = FALSE, randSeed = 0, outputIncrement = 4,
  recordIncrement = 0.1, timeIncrement = 0.001,
  cellTypes = c(new("CellType", name = "DEFAULT")), cellTypeInitFreq = c(1),
  ...)
}
\description{
Constructor for CellModel
}

