% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellSimulation.R
\name{inSilicoCellModel}
\alias{inSilicoCellModel}
\title{Simulates Cell Model}
\usage{
inSilicoCellModel(initialNum, runTime, density,
  modelType = "DrasdoHohme", ...)
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation runs in real cellular time (hours)}

\item{density}{initial density of cell population}

\item{modelType}{the name of the cell-based model to use}

\item{...}{model specific parameters (depends on modelType)}
}
\value{
A CellModel containing all info from the model run
}
\description{
Simulates Cell Model
}
\details{
This function provides a centralized R interface to run c++
code for cell-based models implemented in this package. Standard
parameters, as well as model-specific parameters, are passed in to this
function along with a model name. This function then runs the model and
returns a CellModel object containing all of the information from the
model. This object can then be accessed with various functions designed
to interact with the class. To see a list of available functions, there
is a show() command implemented for CellModel objects.
}
\examples{
inSilicoCellModel(initialNum=1, runTime=8, density=0.1)
}
