% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{inSilicoPathways}
\alias{inSilicoPathways}
\title{This data contains the list of gene targets of modeled signaling pathways}
\usage{
inSilicoPathways
}
\value{
List of gene targets for GtoM, GtoS, and Growth factor pathways.
}
\description{
This data contains the list of gene targets of modeled signaling pathways
}
\details{
List of gene targets for GtoM, GtoS, and Growth factor pathways
modeled in CancerInSilico. GtoM represents genes associated with the G to M
checkpoint of the cell cycle collected from PID pathways PI3KCI_AKT, TNF,
TGFBR, and RB1 in MSigDB v5.1. GtoS represented genes associated with the 
G to S checkpoint of the cell cycle collected from E2F target genes in PID
and hallmark genesets from MSigDb v5.1 and exprimentally validated targets
in TRANSFAC professional 2014. Growth represents the targets of transcription
factors resulting from EGFR activition (STAT1, STAT3, STAT5, Elk-1, Myc,
NF-kappaB, and RelA) from the TRANSFAC professional database 2014 as
described in Fertig et al, 2012.
}
\references{
Subramanian, Tamayo, et al, PNAS, 102:155545, 2005; Matys et al,
Nuc Ac Res, 34:D108, 2006; and Fertig et al, BMC Genomics, 13:160, 2012.
}

