% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inSilicoRNASeq.R
\name{inSilicoRNASeq}
\alias{inSilicoRNASeq}
\title{\code{inSilicoRNASeq} Simulate gene expression data that would be generated from a RNA-seq}
\usage{
inSilicoRNASeq(CellModels, pathways = NULL, ReferenceDataSet = NULL,
  fasta = NULL, samplingFrequency = 1, lambda = 1/3, combineFUN = max,
  attrsep = " ", ...)
}
\arguments{
\item{pathways}{Optional; A list of genes associated with pathways, defaulting to gene symbols in the package data object \code{\link{inSilicoPathways}}. If specified by the user, names of pathways should be GtoM for genes associated with the G to M cell cycle transition, GtoS for genes associated with G to S, Prox for genes associated with contact, and Growth for genes associated with growth factor receptor signaling.}

\item{ReferenceDataSet}{Optional; Reference gene expression dataset to use to calculate the gene expression values for genes in the pathway to use as mean values in the simulation. Defaults values randomly selected from an exponential distribution with parameter \code{lambda}. If specified by the user, \code{row.names} of the ReferenceDataset must match gene names in the \code{pathway} argument.}

\item{fasta}{Optional; NULL value will create a matrix of counts with a negative binomial error model. Otherwise, path to FASTA file containing transcripts from which to simulate reads. See details in \code{\link{simulate_experiment_countmat}}.}

\item{samplingFrequency}{Optional; Distance between time points in hours at which gene expression values are simulated. Defaults to every hour.}

\item{lambda}{Optional; Parameter of the exponential distribution used to determine the maximum expression value of each simulated gene in the pathway. Defaults to 1/3. Not used if values are determined from a dataset in \code{ReferenceDataSet}.}

\item{combineFUN}{Optional; One of max, sum, mean, or median to specify the function used to combine expression values from genes shared in multiple pathways. Defaults to maximum value in any condition (max).}

\item{attrsep}{Optional; string separating gene attributes in the fasta file for \code{\link{polyester}}. Defaults to " ".}

\item{...}{other input arguments to combineFUN and \code{\link{polyester}}}

\item{model}{A \code{\link{CellModel}}}
}
\value{
Matrix simulating the gene expression values for pathway genes from a \code{\link{CellModel}} simulation at the specified frequency in \code{samplingFrequency} if fasta is the default value of NULL. Otherwise, fastq files will be generated to simulate RNA-seq data as described in \code{\link{simulate_experiment_countmat}}.
}
\description{
\code{inSilicoRNASeq} Simulate gene expression data that would be generated from a RNA-seq
}

