% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R
\docType{methods}
\name{interactivePlot}
\alias{interactivePlot}
\alias{interactivePlot,CellModel-method}
\title{plot the cell population and interactively scroll through time points}
\usage{
interactivePlot(model)

\S4method{interactivePlot}{CellModel}(model)
}
\arguments{
\item{model}{cell model object}
}
\value{
plot
}
\description{
plot the cell population and interactively scroll through time points
}

