% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{interactivePlot}
\alias{interactivePlot}
\title{\code{interactivePlot} plots a CellModel and allows the user to control the plot with various commands}
\usage{
interactivePlot(model, time = 0)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
plot a visual representation of cells that takes in command-line-like inputs, type 'h' for help and a list of all available commands
}
\description{
\code{interactivePlot} plots a CellModel and allows the user to control the plot with various commands
}

