% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OffLatticeModel.R
\docType{class}
\name{OffLatticeModel-class}
\alias{OffLatticeModel-class}
\title{OffLatticeModel}
\description{
General description of an off-lattice cell-based model.
not quite a full implementation, but contains much of the neccesary
structure for models of this type
}
\section{Slots}{

\describe{
\item{\code{maxTranslation}}{the largest distance the center of a cell can move}

\item{\code{maxRotation}}{the largest angle a cell can rotate}
}}

