% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{padExpressionMatrix}
\alias{padExpressionMatrix}
\title{Add Noise to Gene Expression Matrices}
\usage{
padExpressionMatrix(mat, nDummyGenes, distr)
}
\arguments{
\item{mat}{matrix of gene expression values}

\item{nGenes}{final number of genes in the expression matrix}

\item{dist}{distribution to sample random values from}
}
\value{
gene expression matrix with dummy genes added in
}
\description{
add genes with random expression values to the matrix
}

