% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Pathway.R
\docType{class}
\name{Pathway-class}
\alias{Pathway-class}
\alias{initialize,Pathway-method}
\title{Pathway Class}
\usage{
\S4method{initialize}{Pathway}(.Object, ...)
}
\arguments{
\item{.Object}{Pathway object}

\item{...}{extra arguments for constructor}
}
\value{
constructed object
}
\description{
Describes the basic properties of a gene pathway

Pathway Class Constructor
}
\details{
expressionScale is a function that accepts three arguments:
model, cell, and time. It should return a number in [0,1] that describes
how active the genes are in this pathway for a given cell in the model
at a given time. In bulk data, the pathway activity is averaged
and transformed by 1 / (1 + exp(-k * (x - M))) where 
k = transformSlope and M = transformMidpoint.
The scale determines how expressed genes in this pathway are. i.e.
and scale of 0 means all genes will have minExpression value and a scale
of 1 means all genes will have maxExpression value. In between these
values the gene expression scales linearly.
}
\section{Slots}{

\describe{
\item{\code{genes}}{names of genes in the pathway}

\item{\code{expressionScale}}{function descibing how this pathway is affected
by the state of the model}

\item{\code{minExpression}}{minimum expression value for each gene (vector)}

\item{\code{maxExpression}}{maximum expression value for each gene (vector)}

\item{\code{transformSlope}}{parameter for transforming bulk data}

\item{\code{transformMidpoint}}{parameter for transforming bulk data}
}}

