% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-OffLatticeModel.R
\docType{methods}
\name{plotCells}
\alias{plotCells}
\alias{plotCells,OffLatticeModel-method}
\title{plot cell population at a given time}
\usage{
plotCells(model, time)

\S4method{plotCells}{OffLatticeModel}(model, time)
}
\arguments{
\item{model}{cell model object}

\item{time}{hour of the model to query}
}
\value{
plot
}
\description{
plot cell population at a given time
}
\examples{
data(SampleModels)
plotCells(modDefault, modDefault@runTime)
}

