% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{plotCells}
\alias{plotCells}
\title{\code{plotCell} plots a CellModel at a given time}
\usage{
plotCells(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
plot a visual representation of cells
}
\description{
\code{plotCell} plots a CellModel at a given time
}

