% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsAtTime.R
\docType{methods}
\name{plotCellsAtTime,CellMatrix-method}
\alias{plotCellsAtTime,CellMatrix-method}
\title{\code{plotCellsAtTime} Plots a CellMatrix at a certain point in time}
\usage{
\S4method{plotCellsAtTime}{CellMatrix}(mat, time)
}
\arguments{
\item{mat}{A Cell Model object}

\item{time}{The timestep at which to plot the matrix. Must be below
the specified max amount of timesteps}
}
\value{
Plot a visual representation of cells at time
}
\description{
\code{plotCellsAtTime} Plots a CellMatrix at a certain point in time
}

