% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsAtTime.R
\docType{methods}
\name{plotCellsAtTime,CellModel-method}
\alias{plotCellsAtTime,CellModel-method}
\title{\code{plotCellsAtTime} Plots a CellModel at a certain point in time}
\usage{
\S4method{plotCellsAtTime}{CellModel}(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{The timestep at which to plot the modelrix.}
}
\value{
Plot a visual representation of cells at time
}
\description{
\code{plotCellsAtTime} Plots a CellModel at a certain point in time
}

