% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrowthRateDistribution.R
\docType{methods}
\name{plotGrowthRateDistribution,CellMatrix-method}
\alias{plotGrowthRateDistribution,CellMatrix-method}
\title{\code{plotGrowthRateDistribution} Plots a distribution of the growth
     rates of all the cells}
\usage{
\S4method{plotGrowthRateDistribution}{CellMatrix}(mat, time)
}
\arguments{
\item{mat}{A Cell Matrix}

\item{time}{The time of interest for the growth rates}
}
\value{
Plots the distribution of growth rates at time
}
\description{
\code{plotGrowthRateDistribution} Plots a distribution of the growth
     rates of all the cells
}

