% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrowthRateDistribution.R
\name{plotGrowthRateDistribution}
\alias{plotGrowthRateDistribution}
\title{\code{plotGrowthRateDistribution} Plots a distribution of the growth rates
of all the cells}
\usage{
plotGrowthRateDistribution(mat, time)
}
\arguments{
\item{mat}{A Cell Matrix}

\item{time}{The time of interest for the growth rates}
}
\description{
\code{plotGrowthRateDistribution} Plots a distribution of the growth rates
of all the cells
}

