% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInteractive.R
\name{plotInteractive}
\alias{plotInteractive}
\title{\code{plotInteractive} Plots a CellModel at a certain point in time}
\usage{
plotInteractive(model, time = 1)
}
\arguments{
<<<<<<< HEAD
\item{model}{A CellModel}
=======
\item{mat}{A CellModel object}
>>>>>>> c9cea553bae792e710af1ccbef8b1ac9c7754d26

\item{time}{The timestep at which to plot the matrix. Must be below
the specified max amount of timesteps}
}
\value{
Plot a visual representation of cells that 
takes in command-line-like inputs as well
see "plotCellsAtTime and plotInteractive" in vignette for
command prompt inputs.
}
\description{
\code{plotInteractive} Plots a CellModel at a certain point in time
}

