% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalCells.R
\docType{methods}
\name{plotTotalCells,CellMatrix-method}
\alias{plotTotalCells,CellMatrix-method}
\title{\code{plotTotalCells} Plots the total number of cells vs time}
\usage{
\S4method{plotTotalCells}{CellMatrix}(mat)
}
\arguments{
\item{mat}{A Cell Matrix}
}
\value{
Plots the size of the cell population over time
}
\description{
\code{plotTotalCells} Plots the total number of cells vs time
}

