% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalCells.R
\docType{methods}
\name{plotTotalCells,CellModel-method}
\alias{plotTotalCells,CellModel-method}
\title{\code{plotTotalCells} Plots the total number of cells vs time}
\usage{
\S4method{plotTotalCells}{CellModel}(model)
}
\arguments{
\item{model}{A CellModel}
}
\value{
Plots the size of the cell population over time
}
\description{
\code{plotTotalCells} Plots the total number of cells vs time
}
\examples{
plotTotalCells(runModel(10,100))
}

