% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalCells.R
\name{plotTotalCells}
\alias{plotTotalCells}
\title{\code{plotTotalCells} Plots the total number of cells vs time}
\usage{
plotTotalCells(model)
}
\arguments{
<<<<<<< HEAD
\item{model}{A CellModel}
=======
\item{mat}{A CellModel}
>>>>>>> c9cea553bae792e710af1ccbef8b1ac9c7754d26
}
\value{
plots the total number of cells over time
}
\description{
\code{plotTotalCells} Plots the total number of cells vs time
}
\examples{
plotTotalCells(runModel(10,100))
}

