% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{printCellTypes}
\alias{printCellTypes}
\title{\code{getCellTypesTable} returns table of cell types in the model}
\usage{
printCellTypes(model, time)
}
\arguments{
\item{model}{a CellModel object}
}
\value{
a table of the cell types
}
\description{
\code{getCellTypesTable} returns table of cell types in the model
}
\examples{
printCellTypes(runCancerSim(1,1), 20)
}

