% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{referenceGeneExpression}
\alias{referenceGeneExpression}
\title{gene expression data for genes in sample pathways}
\usage{
referenceGeneExpression
}
\description{
gene expression data for genes in sample pathways
}
