% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellModel.R, R/class-DrasdoHohmeModel.R
\docType{methods}
\name{run}
\alias{run}
\alias{run,DrasdoHohmeModel-method}
\title{run a cell model}
\usage{
run(model)

\S4method{run}{DrasdoHohmeModel}(model)
}
\arguments{
\item{model}{cell model object}
}
\value{
cell model object with simulation info
}
\description{
run a cell model
}
\examples{
data(SampleModels)
run(modDefault)
}

