% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerModels.R
\name{runDrasdoHohme}
\alias{runDrasdoHohme}
\title{\code{runDrasdoHohme} runs the model based on Drasdo and Hohme (2003)}
\usage{
runDrasdoHohme(initialNum, runTime, density, cycleLengthDist, inheritGrowth,
  outputIncrement, randSeed, ...)
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation represents in realtime}

\item{density}{the density the cells are seeded at}

\item{cycleLengthDist}{cycle time distribution}

\item{inheritGrowth}{whether or not daughter cells have the same cycle-length as parents}

\item{outputIncrement}{time increment to print status at}

\item{randSeed}{seed for the model}

\item{...}{nG, epsilon parameters (specific to this model)}
}
\value{
A CellModel containing all info from the model run
}
\description{
\code{runDrasdoHohme} runs the model based on Drasdo and Hohme (2003)
}
\details{
This function calls the C++ implementation of the Drasdo and Hohme (2003) model.
}

