% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\docType{methods}
\name{show,CellModel-method}
\alias{show,CellModel-method}
\title{\code{show} display summary of CellModel class}
\usage{
\S4method{show}{CellModel}(object)
}
\arguments{
\item{object}{A CellModel Object}
}
\value{
shows all available functions and parameters of model
}
\description{
\code{show} display summary of CellModel class
}
\examples{
show(runCancerSim(1,1))
}
