% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateDrugEffect.R
\name{simulateDrugEffect}
\alias{simulateDrugEffect}
\title{\code{simulateDrugEffect} Simulate Gene Expression Data (Per Cell)}
\usage{
simulateDrugEffect(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A list of pathways, Format:(GtoM, GtoS, Prox)}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateDrugEffect} Simulate Gene Expression Data (Per Cell)
}

