% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{simulateError}
\alias{simulateError}
\title{add simulated error to expression data}
\usage{
simulateError(meanExp, perError, RNAseq = FALSE, singleCell = FALSE,
  polyester = FALSE, fasta = NULL)
}
\arguments{
\item{meanExp}{matrix of gene expression data}

\item{perError}{TODO}

\item{RNAseq}{whether this data is RNA-seq or microarray}
}
\value{
gene expression matrix with error
}
\description{
add noise to all values in expression matrix
}

