% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExp.R, R/simulateGeneExpExt.R
\name{simulateGeneExp}
\alias{simulateGeneExp}
\title{\code{simulateGeneExp} Simulate Gene Expression Data (Beginning)}
\usage{
simulateGeneExp(model, pathway)

simulateGeneExp(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A list of pathways, Format:(GtoM, GtoS, Prox)}

\item{genes}{A vector of gene names}

\item{model}{A CellModel}
}
\value{
the size of the cell population over time

the size of the cell population over time
}
\description{
\code{simulateGeneExp} Simulate Gene Expression Data (Beginning)

\code{simulateGeneExp} Simulate Gene Expression Data (Beginning)
}
\examples{
getTotalCells(runModel(10,100))
getTotalCells(runModel(10,100))
}

