% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExpExt.R
\name{simulateGeneExpExt}
\alias{simulateGeneExpExt}
\title{\code{simulateGeneExpExt} Simulate Gene Expression Data (Beginning)}
\usage{
simulateGeneExpExt(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A list of pathways, Format:(GtoM, GtoS, Prox)}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGeneExpExt} Simulate Gene Expression Data (Beginning)
}
\examples{
getTotalCells(runModel(10,100))
}

