% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExpGroup.R
\name{simulateGeneExpGroup}
\alias{simulateGeneExpGroup}
\title{\code{simulateGeneExpGroup} Simulate Gene Expression Data (Averages)}
\usage{
simulateGeneExpGroup(model, pathway, perError = 0.1, opt = 1)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A list of pathways, Format:(GtoM, GtoS, Prox)}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGeneExpGroup} Simulate Gene Expression Data (Averages)
}

