% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExpSing.R
\name{simulateGeneExpSing}
\alias{simulateGeneExpSing}
\title{\code{simulateGeneExpGroup} Simulate Gene Expression Data (Each Cell)}
\usage{
simulateGeneExpSing(model, pathway, time)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A list of pathways, Format:(GtoM, GtoS, Prox)}

\item{time}{Time Specified by user}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGeneExpGroup} Simulate Gene Expression Data (Each Cell)
}

