% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGrowthFactor.R
\name{simulateGrowthFactor}
\alias{simulateGrowthFactor}
\title{\code{simulateGrowthFactor} Simulate Gene Expression Data (Average)}
\usage{
simulateGrowthFactor(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A vector of gene names}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGrowthFactor} Simulate Gene Expression Data (Average)
}

