% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGToMPath.R
\name{simulateGToMPath}
\alias{simulateGToMPath}
\title{\code{simulateGToMPath} Simulate G2 to Mitosis Phase Gene Expression}
\usage{
simulateGToMPath(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGToMPath} Simulate G2 to Mitosis Phase Gene Expression
}

