% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGToSPath.R
\name{simulateGToSPath}
\alias{simulateGToSPath}
\title{\code{simulateGToSPath} Simulate G1 to Synthesis Phase Gene Expression}
\usage{
simulateGToSPath(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGToSPath} Simulate G1 to Synthesis Phase Gene Expression
}

