% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGtoSPathGroup.R
\name{simulateGtoSPathGroup}
\alias{simulateGtoSPathGroup}
\title{\code{simulateGtoSPathGroup} Simulate G1 to Synthesis Phase Gene Expression (Average)}
\usage{
simulateGtoSPathGroup(model, pathway, samplingFrequency)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}

\item{samplingFrequency}{time (in hours) to sample gene expression data}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGtoSPathGroup} Simulate G1 to Synthesis Phase Gene Expression (Average)
}

