% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMeanExpression.R
\name{simulateMeanSCExpression}
\alias{simulateMeanSCExpression}
\title{\code{simulateSCMeanExpression} Standardized function to simulate mean gene expression values for all single cell RNA-seq}
\usage{
simulateMeanSCExpression(CellModels, ncells = 96, pathways = NULL,
  lambda = 1/3, ReferenceDataSet = NULL, samplingFrequency = 1,
  combineFUN = max, ...)
}
\arguments{
\item{CellModels}{A \code{\link{CellModel}}}

\item{ncells}{Optional; number of cells for single cell RNA-seq. Default=96.}

\item{pathways}{Optional; A list of genes associated with pathways, defaulting to gene symbols in the package data object \code{\link{inSilicoPathways}}. If specified by the user, names of pathways should be GtoM for genes associated with the G to M cell cycle transition, GtoS for genes associated with G to S, Prox for genes associated with contact, and Growth for genes associated with growth factor receptor signaling.}

\item{lambda}{Optional; Parameter of the exponential distribution used to determine the maximum expression value of each simulated gene in the pathway. Defaults to 1/3. Not used if values are determined from a dataset in \code{ReferenceDataSet}.}

\item{ReferenceDataSet}{Optional; Reference gene expression dataset to use to calculate the gene expression values for genes in the pathway to use as mean values in the simulation. Defaults values randomly selected from an exponential distribution with parameter \code{lambda}. If specified by the user, \code{row.names} of the ReferenceDataset must match gene names in the \code{pathway} argument.}

\item{samplingFrequency}{Optional; Distance between time points in hours at which gene expression values are simulated. Defaults to every hour.}

\item{combineFUN}{Optional; Function used to combine expression values from genes shared in multiple pathways. Defaults to maximum value in any condition.}

\item{...}{other input arguments to combineFUN}
}
\value{
Matrix simulating the mean gene expression values for pathway genes from a \code{\link{CellModel}} simulation at the specified frequency in \code{samplingFrequency}.
}
\description{
\code{simulateSCMeanExpression} Standardized function to simulate mean gene expression values for all single cell RNA-seq
}

