% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{simulatePathway}
\alias{simulatePathway}
\title{\code{simulatePathway} Simulates gene expression data for a given pathway}
\usage{
simulatePathway(model, pathway, type, sampFreq = 1, sampSize = 1,
  singleCell = FALSE)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{The pathway to simulate expression for}

\item{type}{the type of cellular function the pathway is related too}

\item{sampFreq}{how often to sample the cells and record gene expression data}

\item{sampSize}{how many cells to sample in the case of single-cell expression data}

\item{singleCell}{whether or not to simulate single-cell expression data}
}
\value{
gene expression matrix for given pathway
}
\description{
\code{simulatePathway} Simulates gene expression data for a given pathway
}
