% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Pathway.R
\name{simulatePathwayActivity}
\alias{simulatePathwayActivity}
\title{simulate pathway activity based in model behavior}
\usage{
simulatePathwayActivity(pathway, model, sampFreq, sampSize,
  singleCell = FALSE)
}
\arguments{
\item{pathway}{a 'Pathway' object}

\item{model}{a 'CellModel' object}

\item{sampFreq}{time in between samples}

\item{singleCell}{simulate single cell activity}

\item{sampleSize}{number of cells to sample in the case of single cell}
}
\value{
vector with values in [0,1] indicatin pathway activity
}
\description{
simulates pathway activity over the course of a cell model
}
\keyword{internal}

