% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Pathway.R
\name{simulatePathwayExpression}
\alias{simulatePathwayExpression}
\title{Simulate Mean Expression for a Pathway}
\usage{
simulatePathwayExpression(pathway, model, sampFreq, singleCell = FALSE,
  sampSize = 0)
}
\arguments{
\item{pathway}{a 'Pathway' object}

\item{model}{a 'CellModel' object}

\item{sampFreq}{time in between samples of gene expression}

\item{singleCell}{simulate single cell data}

\item{sampleSize}{number of cells to sample in the case of single cell}
}
\value{
matrix of gene expression data for this pathway
}
\description{
simulates gene expression data for a single pathway,
 does not incorporate an error model
}
\keyword{internal}

