% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateProxPath.R
\name{simulateProxPath}
\alias{simulateProxPath}
\title{\code{simulateProxPath} Simulate Proximity Based Gene Expression}
\usage{
simulateProxPath(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateProxPath} Simulate Proximity Based Gene Expression
}

