% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateProxPathSing.R
\name{simulateProxPathSing}
\alias{simulateProxPathSing}
\title{\code{simulateProxPathSing} Simulate Proximity Based Gene Expression (Per Cell)}
\usage{
simulateProxPathSing(model, pathway, sampFreq = 1, ncell = 0)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}

\item{sampFreq}{Time (in hours) at which to simulate gene expression data}

\item{nCell}{Number of cells selected for a random sample at given time}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateProxPathSing} Simulate Proximity Based Gene Expression (Per Cell)
}

