% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{simulateWithLimmaVoom}
\alias{simulateWithLimmaVoom}
\title{Bulk RNA-seq Error Model}
\usage{
simulateWithLimmaVoom(meanExp, params)
}
\arguments{
\item{meanExp}{matrix of mean expression values}

\item{params}{GenExpressionParams object}
}
\value{
matrix with error model incorporated
}
\description{
adds the limma-voom error model to the data
}
\keyword{internal}

