% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{simulateWithSplatter}
\alias{simulateWithSplatter}
\title{Single Cell RNA-seq Error Model}
\usage{
simulateWithSplatter(meanExp, params)
}
\arguments{
\item{meanExp}{matrix of mean expression values}

\item{params}{GenExpressionParams object}
}
\value{
matrix with error model incorporated
}
\description{
adds both the limma-voom error model and calculates dropout
}
\keyword{internal}
