% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitCells.R
\name{splitCells}
\alias{splitCells}
\title{\code{splitCells} Determines which cells are in the process of dividing}
\usage{
splitCells(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{Time}
}
\value{
the size of the cell population over time
}
\description{
\code{splitCells} Determines which cells are in the process of dividing
}

