% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{timeToRow}
\alias{timeToRow}
\title{\code{timeToRow} return the correct row in the mCells list corresponding to a given time}
\usage{
timeToRow(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
corresponding row in mCells list
}
\description{
\code{timeToRow} return the correct row in the mCells list corresponding to a given time
}

