% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{voomErrorModel}
\alias{voomErrorModel}
\title{Error Model found in Limma-Voom}
\usage{
voomErrorModel(mu, bcvCommon, bcvDF)
}
\arguments{
\item{mu}{mean expression matrix}

\item{bcvCommon}{biological variation parameter}

\item{bcvDF}{degrees of freedom for the chisq distribution}
}
\value{
list of the cell means and the simulated counts
}
\description{
Core error model presented in original voom paper, used for 
multiple RNA-seq simulations
}
\keyword{internal}
