% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Pathway.R
\name{calibratePathway}
\alias{calibratePathway}
\title{calibrate pathway with data}
\usage{
calibratePathway(pathway, dataSet)
}
\arguments{
\item{pathway}{a 'Pathway' object}

\item{dataSet}{reference data set}
}
\value{
pathway with min/max values for expression based on data set
}
\description{
sets the min and max values for each gene in a pathway
based on a data set
}
