% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R, R/getCycleLengthDistribution.R, R/getDensity.R, R/getNumberOfCells.R, R/getParameters.R, R/plotCellsAtTime.R, R/plotInteractive.R
\docType{class}
\name{CellModel-class}
\alias{CellModel-class}
\alias{getCycleLengthDistribution,CellModel-method}
\alias{getDensity,CellModel-method}
\alias{getNumberOfCells,CellModel-method}
\alias{getParameters,CellModel-method}
\alias{plotCellsAtTime,CellModel-method}
\alias{plotInteractive,CellModel-method}
\alias{show,CellModel-method}
\title{An S4 class to represent the output of a cell-based model}
\usage{
\S4method{show}{CellModel}(object)

\S4method{getCycleLengthDistribution}{CellModel}(model, time)

\S4method{getDensity}{CellModel}(model, time)

\S4method{getNumberOfCells}{CellModel}(model, time)

\S4method{getParameters}{CellModel}(model, fullDist = FALSE)

\S4method{plotCellsAtTime}{CellModel}(model, time)

\S4method{plotInteractive}{CellModel}(model, time = 0)
}
\arguments{
\item{object}{A CellModel Object}

\item{model}{A CellModel}

\item{time}{The time of interest for the growth rates}

\item{fullDist}{whether or not to return full distribution of cycle length}

\item{model}{A Cell Model}

\item{time}{The time of interest}

\item{model}{A CellModel}

\item{time}{the time of interest}

\item{model}{A CellModel}

\item{model}{A CellModel}

\item{time}{The timestep at which to plot the modelrix.}

\item{model}{A CellModel}

\item{time}{The timestep at which to plot the matrix. Must be below
the specified max amount of timesteps}
}
\value{
returns a vector of growth rates for cells alive at time

The density of cells at that time (not quite the same as confluency)

the size of the cell population over time

a named list of parameters in the model

Plot a visual representation of cells at time

Plot a visual representation of cells that 
takes in command-line-like inputs as well
see "plotCellsAtTime and plotInteractive" in vignette for
command prompt inputs.
}
\description{
An S4 class to represent the output of a cell-based model

\code{show} display summary of CellModel class

\code{getCycleLengthDistribution} Plots a distribution of the growth
     rates of all the cells

\code{getDensity} Gets the density of cells at a given time

\code{getNumberOfCells} Plots the total number of cells vs time

\code{getParameters} get a named list of parameters in the model

\code{plotCellsAtTime} Plots a CellModel at a certain point in time

\code{plotInteractive} Plots a CellModel at a certain point in time
}
\section{Slots}{

\describe{
\item{\code{m_cells}}{A list object where each row of the list describes the state
of all the cells in the model at a given time. Each cell is described over
6 columns: [1] x-coordinate, [2] y-coordinate, [3] radius, [4] axis length,
[5] axis angle, [6] growth rate. For instance, the x-coordinates of the first
3 cells will be in columns 1,7,13.}

\item{\code{m_initialNumCells}}{the initial number of cells in the model}

\item{\code{m_runTime}}{the total run time (hours) of the model}

\item{\code{m_initialDensity}}{the density the cells were seeded at}

\item{\code{m_inheritGrowth}}{whether or not cells inherit growth rates from their parent}

\item{\code{m_outputIncrement}}{the frequency of print statements during the run}

\item{\code{m_randSeed}}{the random seed}

\item{\code{m_epsilon}}{model specific parameter}

\item{\code{m_nG}}{model specific parameter}

\item{\code{m_timeIncrement}}{amount of time elapsed in each model step}

\item{\code{m_cycleLengthDist}}{initial distribution of cell-cycle lengths}
}}
\examples{
show(runCancerSim(1,6))
getCycleLengthDistribution(runCancerSim(1,10),0)
getDensity(runCancerSim(1,10),10)
getNumberOfCells(runCancerSim(1,10),10)
getParameters(runCancerSim(1,1))
plotCellsAtTime(runCancerSim(1,10),10)
}

