% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\docType{class}
\name{CellModel-class}
\alias{CellModel-class}
\title{CellModel}
\description{
An S4 class that contains the output of a cell simulation,
     along with all the parameters used to generate the simulation
}
\section{Slots}{

\describe{
\item{\code{mCells}}{A list object where each row of the list describes the
state of all the cells in the model at a given time. Each cell is
described over 6 columns: [1] x-coordinate, [2] y-coordinate, [3]
radius, [4] axis length, [5] axis angle, [6] growth rate. For
instance, the x-coordinates of the first 3 cells will be in columns
1,7,13.}

\item{\code{mParameters}}{A list object containing all parameters used to run
the model}
}}

