% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\docType{class}
\name{CellModel-class}
\alias{CellModel-class}
\title{CellModel}
\description{
An S4 class that contains the output of a cell simulation,
 along with all the parameters used to generate the simulation
}
\section{Slots}{

\describe{
\item{\code{cells}}{A list object that describes the state of the cells
at each time. The state representation depends on the type of
model run, and is accessed by the function designed for each
model type.}

\item{\code{params}}{A list object containing all parameters used in the model}
}}

