% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R, R/ExternalFunctions.R
\docType{class}
\name{CellModel-class}
\alias{CellModel-class}
\alias{getParameters}
\alias{interactivePlot}
\alias{show,CellModel-method}
\title{An S4 class to represent the output of a cell-based model}
\usage{
\S4method{show}{CellModel}(object)

getParameters(model, fullDist = FALSE)

interactivePlot(model, time = 0)
}
\arguments{
\item{object}{A CellModel Object}

\item{model}{A CellModel}

\item{fullDist}{whether or not to return full distribution of cycle length}

\item{time}{The timestep at which to plot the matrix. Must be below
the specified max amount of timesteps}

\item{model}{A CellModel}
}
\value{
a named list of parameters in the model

Plot a visual representation of cells that 
takes in command-line-like inputs as well
see "plotCellsAtTime and plotInteractive" in vignette for
command prompt inputs.
}
\description{
An S4 class to represent the output of a cell-based model

\code{show} display summary of CellModel class

\code{getParameters} get a named list of parameters in the model

\code{interactivePlot} Plots a CellModel and allows the user to control the plot with various commands
}
\section{Slots}{

\describe{
\item{\code{m_cells}}{A list object where each row of the list describes the state
of all the cells in the model at a given time. Each cell is described over
6 columns: [1] x-coordinate, [2] y-coordinate, [3] radius, [4] axis length,
[5] axis angle, [6] growth rate. For instance, the x-coordinates of the first
3 cells will be in columns 1,7,13.}

\item{\code{m_initialNumCells}}{the initial number of cells in the model}

\item{\code{m_runTime}}{the total run time (hours) of the model}

\item{\code{m_initialDensity}}{the density the cells were seeded at}

\item{\code{m_inheritGrowth}}{whether or not cells inherit growth rates from their parent}

\item{\code{m_outputIncrement}}{the frequency of print statements during the run}

\item{\code{m_randSeed}}{the random seed}

\item{\code{m_epsilon}}{model specific parameter}

\item{\code{m_nG}}{model specific parameter}

\item{\code{m_timeIncrement}}{amount of time elapsed in each model step}

\item{\code{m_cycleLengthDist}}{initial distribution of cell-cycle lengths}
}}
\examples{
show(runCancerSim(1,6))
getParameters(runCancerSim(1,1))
}

