% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType-class.R
\docType{class}
\name{CellType-class}
\alias{CellType-class}
\title{CellType}
\description{
An S4 class to represent the type of cell (and its consequent behavior)
}
\section{Slots}{

\describe{
\item{\code{mType}}{the type of cell as a character}
}}

