% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CellType.R
\docType{class}
\name{CellType-class}
\alias{CellType-class}
\title{CellType}
\description{
The properties of a cell type
}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the cell type}

\item{\code{size}}{the relative size (volume) of the cell}

\item{\code{minCycle}}{minimum possible cell cycle length}

\item{\code{cycleLength}}{function that returns sample from distribution of 
cycle lengths}
}}

