% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType-class.R
\docType{class}
\name{CellType-class}
\alias{CellType-class}
\title{CellType}
\description{
An S4 class to the properties of a cell type
}
\section{Slots}{

\describe{
\item{\code{type}}{the name of the cell type}

\item{\code{size}}{the relative size (volume) of the cell}

\item{\code{cycleLenth}}{function that returns the average (target) cycle
length of this cell}

\item{\code{drugEffect}}{function that returns the effect a drug has on the
the cell properties}

\item{\code{inheritCycleLength}}{bool - inherit cycle length from parent}

\item{\code{inheritDrugEffect}}{bool - inherit drug effect from parent}
}}

