% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType-class.R
\docType{class}
\name{CellType-class}
\alias{CellType-class}
\title{CellType}
\description{
An S4 class to the properties of a cell type
}
\section{Slots}{

\describe{
\item{\code{mType}}{the name of the cell type}

\item{\code{mCycleLenth}}{distribution of cycle lengths for this cell}

\item{\code{mDrugEffect}}{function descibing effect of a drug}
}}

