% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-GeneExpressionParams.R
\docType{class}
\name{GeneExpressionParams-class}
\alias{GeneExpressionParams-class}
\title{GeneExpressionParams}
\description{
Parameters for simulating gene expression
}
\section{Slots}{

\describe{
\item{\code{sampFreq}}{how often to generate data}

\item{\code{nGenes}}{total number of genes used (matrix padded with dummys)}

\item{\code{combineFUN}}{function used to combine gene expression data}

\item{\code{singleCell}}{logical: simulate single cell data}

\item{\code{nCells}}{number of cells to use for single cell data}

\item{\code{perError}}{percent error?}

\item{\code{microArray}}{true if micro array data}

\item{\code{randSeed}}{random seed for simulation}
}}

