% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellPhasePos.R
\name{getCellPhasePos}
\alias{getCellPhasePos}
\title{\code{getCellPhasePos} Determines which cells are in the process of dividing}
\usage{
getCellPhasePos(model, time, sampFreq)
}
\arguments{
\item{model}{A CellModel}

\item{time}{Time}

\item{sampFreq}{Time (in hours) at which to simulate gene expression data}
}
\value{
the size of the cell population over time
}
\description{
\code{getCellPhasePos} Determines which cells are in the process of dividing
}

