% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getNumberOfCells}
\alias{getNumberOfCells}
\title{\code{getNumberOfCells} get the number of cells alive}
\usage{
getNumberOfCells(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{time in model hours}
}
\value{
the number of cells at this time
}
\description{
\code{getNumberOfCells} get the number of cells alive
}
\examples{
getNumberOfCells(runCancerSim(1,1), 1)
}
