% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\docType{methods}
\name{getParameters,CellModel-method}
\alias{getParameters,CellModel-method}
\title{\code{getParameters} get a named list of parameters in the model}
\usage{
\S4method{getParameters}{CellModel}(model, fullDist = FALSE)
}
\arguments{
\item{model}{A CellModel}

\item{fullDist}{whether or not to return full distribution of cycle length}
}
\value{
a named list of parameters in the model
}
\description{
\code{getParameters} get a named list of parameters in the model
}
\examples{
getParameters(runModel(1,1))
}

