% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellModel-class.R
\name{getParameters}
\alias{getParameters}
\title{\code{getParameters} get a named list of parameters in the model}
\usage{
getParameters(model)
}
\arguments{
\item{model}{A CellModel}

\item{fullDist}{[bool] return full distribution of cycle length}
}
\value{
a named list of parameters in the model
}
\description{
\code{getParameters} get a named list of parameters in the model
}
\examples{
getParameters(runCancerSim(1,1))
}

