% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-OffLatticeModel.R
\docType{methods}
\name{initialize,OffLatticeModel-method}
\alias{initialize,OffLatticeModel-method}
\title{Off-Lattice Model Constructor}
\usage{
\S4method{initialize}{OffLatticeModel}(.Object, maxTranslation = 0.1,
  maxRotation = 0.3, ...)
}
\arguments{
\item{.Object}{OffLatticeModel object}

\item{maxTranslation}{maximum movement of cell}

\item{maxRotation}{maximim rotation of mitosis cell}

\item{...}{model specific parameters}
}
\value{
initialized cell model
}
\description{
Off-Lattice Model Constructor
}

