% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{inSilicoGeneExpression}
\alias{inSilicoGeneExpression}
\title{simulate gene expression data}
\usage{
inSilicoGeneExpression(model, pathways,
  params = new("GeneExpressionParams"))
}
\arguments{
\item{model}{a CellModel object}

\item{pathways}{list of genes pathways}

\item{params}{GeneExpressionParams object}
}
\value{
list of pathway activity and gene expression
}
\description{
simulate gene expression data for a set of pathways, using
the behavior of a CellModel as the basis for the simulation
}
