% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{negBinError}
\alias{negBinError}
\title{Negative Binomial Error model}
\usage{
negBinError(meanExp, dataSet = NULL, invChisq = TRUE)
}
\arguments{
\item{meanExp}{matrix of mean expression data}

\item{dataSet}{reference data set of gene expression}

\item{invChisq}{use chi-squared distribution}
}
\description{
#TODO
}
\keyword{internal}

