% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{negBinError}
\alias{negBinError}
\title{Negative Binomial Error model from limma voom}
\usage{
negBinError(meanExp, invChisq = TRUE)
}
\arguments{
\item{meanExp}{matrix of mean expression data}

\item{invChisq}{use chi-squared distribution}
}
\description{
using the methods in limma-voom paper, simulate final counts
 from matrix of expected counts
}
\keyword{internal}

