% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Pathway.R
\docType{class}
\name{Pathway-class}
\alias{Pathway-class}
\alias{initialize,Pathway-method}
\title{Pathway Class}
\usage{
\S4method{initialize}{Pathway}(.Object, ...)
}
\description{
Describes the basic properties of a gene pathway

Pathway Class Constructor
}
\section{Slots}{

\describe{
\item{\code{genes}}{names of genes in the pathway}

\item{\code{expressionScale}}{function descibing how this pathway is affected
by the state of the model}

\item{\code{minExpression}}{minimum expression value for each gene (vector)}

\item{\code{maxExpression}}{maximum expression value for each gene (vector)}
}}

