% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsAtTime.R
\name{plotCellsAtTime}
\alias{plotCellsAtTime}
\title{\code{plotCellsAtTime} Plots a CellMatrix at a certain point in time}
\usage{
plotCellsAtTime(mat, time)
}
\arguments{
\item{mat}{A Cell Matrix}

\item{time}{The timestep at which to plot the matrix. Must be below
the specified max amount of timesteps}
}
\value{
Plot a visual representation of cells at time
}
\description{
\code{plotCellsAtTime} Plots a CellMatrix at a certain point in time
}
\examples{
plotCellsAtTime(runModel(100,10),5)
}

