% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellsAtTime.R
\name{plotCellsAtTime}
\alias{plotCellsAtTime}
\title{\code{plotCellsAtTime} Plots a CellModel at a certain point in time}
\usage{
plotCellsAtTime(model, time)
}
\arguments{
\item{model}{A CellModel}

\item{time}{The timestep at which to plot the modelrix.}
}
\value{
Plot a visual representation of cells at time
}
\description{
\code{plotCellsAtTime} Plots a CellModel at a certain point in time
}
\examples{
plotCellsAtTime(runModel(1,10),10)
}

