% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTotalCells.R
\name{plotTotalCells}
\alias{plotTotalCells}
\title{\code{plotTotalCells} Plots the total number of cells vs time}
\usage{
plotTotalCells(mat)
}
\arguments{
\item{mat}{A CellModel}
}
\value{
Plots the size of the cell population over time
}
\description{
\code{plotTotalCells} Plots the total number of cells vs time
}
\examples{
plotTotalCells(runModel(10,100))
}

