% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCancerSim.R
\name{runCancerSim}
\alias{runCancerSim}
\title{\code{runCancerSim} Runs the model}
\usage{
runCancerSim(initialNum, runTime, density = 0.01, cycleLengthDist = 12,
  drugEffect = getDrugEffect(cycleLengthDist = cycleLengthDist),
  inheritGrowth = FALSE, outputIncrement = 6, recordIncrement = 0.25,
  randSeed = 0, modelType = "DrasdoHohme2003", ...)
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation represents in realtime}

\item{density}{the density the cells are seeded at}

\item{cycleLengthDist}{cycle time distribution}

\item{drugEffect}{distribution of drug effects}

\item{inheritGrowth}{whether or not daughter cells have the same cycle-length as parents}

\item{outputIncrement}{time increment to print status at}

\item{randSeed}{seed for the model}

\item{modelType}{the name of the cell-based model to use}

\item{...}{model specific parameters (depends on modelType)}
}
\value{
A CellModel containing all info from the model run
}
\description{
\code{runCancerSim} Runs the model
}
\examples{
runCancerSim(1,8)
}

