% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerModels.R
\name{runCancerSim}
\alias{runCancerSim}
\title{\code{runCancerSim} Runs the model}
\usage{
runCancerSim(initialNum, runTime, density = 0.01, cycleLengthDist = 12,
  drugEffect = getDrugEffect(cycleLenthDist = cycleLengthDist),
  inheritGrowth = FALSE, outputIncrement = 6, recordIncrement = 0.25,
  randSeed = 0, modelType = "DrasdoHohme2003", drugTime = 0, ...)
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation runs in real cellular time (hours)}

\item{density}{the density the cells are seeded at}

\item{cycleLengthDist}{cycle time distribution}

\item{drugEffect}{distribution of drug effects}

\item{inheritGrowth}{whether or not daughter cells have the same cycle-length as parents}

\item{outputIncrement}{time increment to print status at}

\item{randSeed}{seed for the model}

\item{modelType}{the name of the cell-based model to use}

\item{...}{model specific parameters (depends on modelType)}
}
\value{
A CellModel containing all info from the model run
}
\description{
\code{runCancerSim} Runs the model
}
\details{
This function provides a centralized R interface to run c++ code for cell-based models implemented in this package. Standard parameters, as well as model-specific parameters, are passed in to this function along with a model name. This function then runs the model and returns a CellModel object containing all of the information from the model. This object can then be accessed with various functions designed to interact with the class. To see a list of available functions, there is a show() command implemented for CellModel objects.
}
\examples{
runCancerSim(1,8)
}

