% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerModels.R
\name{runDrasdoHohme}
\alias{runDrasdoHohme}
\title{\code{runDrasdoHohme} Runs the model from the Drasdo, Hohme paper}
\usage{
runDrasdoHohme(initialNum, runTime, density, cycleLengthDist, inheritGrowth,
  outputIncrement, recordIncrement, randSeed, drugEffect, drugTime, boundary,
  cycleSyncProb, ...)
}
\description{
\code{runDrasdoHohme} Runs the model from the Drasdo, Hohme paper
}
\details{

}

