% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{\code{runModel} Runs the model}
\usage{
runModel(initialNum, runTime, density = 0.05, maxMigration = 0.5,
  maxDeform = 0.075, maxRotate = 0.3, epsilon = 0.05, delta = 5,
  outIncrement = 10, randSeed = 0, growthRates = runif(initialNum, 0.05,
  0.25))
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation runs}

\item{density}{the density the cells are seeded at}

\item{maxMigration}{farthest a cell can move in a single update}

\item{maxDeform}{most a cell can deform in a single update}

\item{maxRotate}{most a cell can rotate in a single update}

\item{epsilon}{model specific parameter}

\item{delta}{model specific parameter}

\item{outIncrement}{time increment to print status at}

\item{randSeed}{seed for the model}

\item{meanGrowth}{mean growth rate of cells}

\item{varGrowth}{the variance of the growth rate (assumed normal)}
}
\value{
A CellModel containing all info from the model run
}
\description{
\code{runModel} Runs the model
}
\examples{
runModel(100,10)
}

