% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{\code{runModel} Runs the model}
\usage{
runModel(initialNum, runTime, density = 0.01, cycleLengthDist = 12,
  inheritGrowth = F, timeIncrement = 0.25, outputIncrement = 40,
  randSeed = 0, epsilon = 4)
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation represents in realtime}

\item{density}{the density the cells are seeded at}

\item{cycleLengthDist}{cycle time distribution}

\item{inheritGrowth}{whether or not daughter cells have the same cycle-length as parents}

\item{timeIncrement}{time length of a step in the model, based on the hour}

\item{outputIncrement}{time increment to print status at}

\item{randSeed}{seed for the model}

\item{epsilon}{epsilon model specific parameter}
}
\value{
A CellModel containing all info from the model run
}
\description{
\code{runModel} Runs the model
}
\examples{
runModel(100,10)
}

