% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{\code{runModel} Runs the model}
\usage{
runModel(initialNum, runTime, density = 0.01, cycleLengthDist = 12,
<<<<<<< HEAD
  inheritGrowth = F, outputIncrement = 6, randSeed = 0, epsilon = 4,
  nG = 10)
=======
  inheritGrowth = FALSE, outputIncrement = 6, randSeed = 0,
  epsilon = 10, nG = 24)
>>>>>>> 5a4936dd8591d8e541975803bf2813ea7474e081
}
\arguments{
\item{initialNum}{how many cells initially}

\item{runTime}{how long the simulation represents in realtime}

\item{density}{the density the cells are seeded at}

\item{cycleLengthDist}{cycle time distribution}

\item{inheritGrowth}{whether or not daughter cells have the same cycle-length as parents}

\item{outputIncrement}{time increment to print status at}

\item{randSeed}{seed for the model}

<<<<<<< HEAD
\item{epsilon}{epsilon model specific parameter}
=======
\item{epsilon}{model specific parameter}
>>>>>>> 5a4936dd8591d8e541975803bf2813ea7474e081

\item{nG}{model specific parameter}
}
\value{
A CellModel containing all info from the model run
}
\description{
\code{runModel} Runs the model
}
\examples{
runModel(1,8)
}

