% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R
\name{simulateError}
\alias{simulateError}
\title{add simulated error to expression data}
\usage{
simulateError(meanExp, dataSet = NULL, perError, microArray)
}
\arguments{
\item{meanExp}{matrix of gene expression data}

\item{dataSet}{matrix of gene expression data where row names are genes}

\item{perError}{TODO}

\item{microArray}{whether this data is RNA-seq or microarray}
}
\value{
gene expression matrix with error
}
\description{
add noise to all values in expression matrix
}

