% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExp.R
\name{simulateGeneExp}
\alias{simulateGeneExp}
\title{\code{getTotalCells} Simulate Gene Expression Data (Beginning)}
\usage{
simulateGeneExp(model, time, genes)
}
\arguments{
\item{model}{A CellModel}
}
\value{
the size of the cell population over time
}
\description{
\code{getTotalCells} Simulate Gene Expression Data (Beginning)
}
\examples{
getTotalCells(runModel(10,100))
}

