% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExpSing.R
\name{simulateGeneExpSing}
\alias{simulateGeneExpSing}
\title{\code{simulateGeneExpSing} Simulate Gene Expression Data (Per Cell)}
\usage{
simulateGeneExpSing(model, pathways, perError = 0.1, opt = 1,
  sampFreq = 1)
}
\arguments{
\item{model}{A CellModel}

\item{pathways}{A list of pathways, Format:(GtoM, GtoS, Prox)}

\item{perError}{User defined error for noise calculations}

\item{opt}{Option for which noise error calculated}

\item{sampFreq}{The frequency at which data is collected}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGeneExpSing} Simulate Gene Expression Data (Per Cell)
}

