% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGeneExpSing.R
\name{simulateGeneExpSing}
\alias{simulateGeneExpSing}
\title{\code{simulateGeneExpSing} Simulate Gene Expression Data (Per Cell)}
\usage{
simulateGeneExpSing(model, pathway, perError = 0.1, opt = 1)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A list of pathways, Format:(GtoM, GtoS, Prox)}

\item{perError}{User defined error for noise calculations}

\item{opt}{Option for which noise error calculated}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGeneExpSing} Simulate Gene Expression Data (Per Cell)
}

