% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGrowthPathGroup.R
\name{simulateGrowthFactor}
\alias{simulateGrowthFactor}
\title{\code{simulateGrowthFactor} Simulate Gene Expression Data (Average)}
\usage{
simulateGrowthFactor(model, pathway, samplingFrequency)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A vector of gene names}

\item{samplingFrequency}{Time (in hours) at which to simulate gene expression data}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGrowthFactor} Simulate Gene Expression Data (Average)
}

