% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGtoMPathGroup.R
\name{simulateGtoMPathGroup}
\alias{simulateGtoMPathGroup}
\title{\code{simulateGtoMPathGroup} Simulate G2 to Mitosis Phase Gene Expression (Average)}
\usage{
simulateGtoMPathGroup(model, pathway, sampFreq = 1)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}

\item{sampFreq}{Time (in hours) at which to simulate gene expression data}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGtoMPathGroup} Simulate G2 to Mitosis Phase Gene Expression (Average)
}

