% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGToMPathGroup.R
\name{simulateGToMPathGroup}
\alias{simulateGToMPathGroup}
\title{\code{simulateGToMPathGroup} Simulate G2 to Mitosis Phase Gene Expression (Average)}
\usage{
simulateGToMPathGroup(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGToMPathGroup} Simulate G2 to Mitosis Phase Gene Expression (Average)
}

