% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGToMPathSing.R
\name{simulateGToMPathSing}
\alias{simulateGToMPathSing}
\title{\code{simulateGToMPathSing} Simulate G2 to Mitosis Phase Gene Expression (Per Cell)}
\usage{
simulateGToMPathSing(model, pathway, sampFreq = 1)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}

\item{sampFreq}{The frequency at which data is collected}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGToMPathSing} Simulate G2 to Mitosis Phase Gene Expression (Per Cell)
}

