% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGtoMPathSing.R
\name{simulateGtoMPathSing}
\alias{simulateGtoMPathSing}
\title{\code{simulateGtoMPathSing} Simulate G2 to Mitosis Phase Gene Expression (Per Cell)}
\usage{
simulateGtoMPathSing(model, pathway, sampFreq = 1, ncell = 0)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}

\item{sampFreq}{Time (in hours) at which to simulate gene expression data}

\item{nCell}{Number of cells selected for a random sample at given time}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGtoMPathSing} Simulate G2 to Mitosis Phase Gene Expression (Per Cell)
}

