% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGToSPathGroup.R
\name{simulateGToSPathGroup}
\alias{simulateGToSPathGroup}
\title{\code{simulateGToSPathGroup} Simulate G1 to Synthesis Phase Gene Expression (Average)}
\usage{
simulateGToSPathGroup(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGToSPathGroup} Simulate G1 to Synthesis Phase Gene Expression (Average)
}

