% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGToSPathSing.R
\name{simulateGToSPathSing}
\alias{simulateGToSPathSing}
\title{\code{simulateGToSPathSing} Simulate G1 to Synthesis Phase Gene Expression (Per Cell)}
\usage{
simulateGToSPathSing(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateGToSPathSing} Simulate G1 to Synthesis Phase Gene Expression (Per Cell)
}

