% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionSimulation.R, R/PathwaySimulation.R
\name{simulatePathway}
\alias{simulatePathway}
\title{\code{simulatePathway} Simulates gene expression data for a given pathway}
\usage{
simulatePathway(model, pathway, type, sampFreq = 1, sampSize,
  singleCell = FALSE, timeWindow = 1, downReg = FALSE)

simulatePathway(model, pathway, type, sampFreq = 1, sampSize,
  singleCell = FALSE, timeWindow = 1, downReg = FALSE)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{The pathway to simulate expression for}

\item{type}{the type of cellular function the pathway is related too}

\item{sampFreq}{how often to sample the cells and record gene expression data}

\item{sampSize}{how many cells to sample in the case of single-cell expression data}

\item{singleCell}{whether or not to simulate single-cell expression data}

\item{timeWindow}{Length of window examined to see if relevant cell
state changed (for 'S' and 'M' pathways)}

\item{downReg}{T/F: pathway is down regulated by cell activity (type)}

\item{model}{A CellModel object}

\item{pathway}{List of genes and range of expression values (fields:
'genes', 'max', 'min')}

\item{type}{Cellular function related to pathway ('S', 'M', 'PROX',
'GROWTH')}

\item{sampFreq}{Time (hrs) between each sample}

\item{sampSize}{Size of sample for single cell data}

\item{singleCell}{T/F: generate single cell data}
}
\value{
gene expression matrix for given pathway

Gene expression matrix for given pathway
}
\description{
\code{simulatePathway} Simulates gene expression data for a given pathway

Simulates gene expression data for a given pathway
}

