% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwaySimulation.R
\name{simulatePathway}
\alias{simulatePathway}
\title{\code{simulatePathway}}
\usage{
simulatePathway(model, pathway, type, sampFreq = 1, sampSize,
  singleCell = FALSE, timeWindow = 1, downReg = FALSE)
}
\arguments{
\item{model}{A CellModel object}

\item{pathway}{List of genes and range of expression values (fields:
'genes', 'max', 'min')}

\item{type}{Cellular function related to pathway ('S', 'M', 'PROX',
'GROWTH')}

\item{sampFreq}{Time (hrs) between each sample}

\item{sampSize}{Size of sample for single cell data}

\item{singleCell}{T/F: generate single cell data}

\item{timeWindow}{Length of window examined to see if relevant cell
state changed (for 'S' and 'M' pathways)}

\item{downReg}{T/F: pathway is down regulated by cell activity (type)}
}
\value{
Gene expression matrix for given pathway
}
\description{
Simulates gene expression data for a given pathway
}
