% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Pathway.R
\name{simulatePathwayExpression}
\alias{simulatePathwayExpression}
\title{simulate gene expression data for a single pathway}
\usage{
simulatePathwayExpression(pathway, activity)
}
\arguments{
\item{pathway}{object of 'Pathway' class}

\item{activity}{named vector of pathway activity in [0,1]}
}
\value{
scale applied to min/max expression values of the pathway
}
\description{
applies given pathway activity to the gene expression
 ranges to produce simulated gene expression data
}
\keyword{internal}

