% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateProxPathGroup.R
\name{simulateProxPathGroup}
\alias{simulateProxPathGroup}
\title{\code{simulateProxPathGroup} Simulate Proximity Based Gene Expression (Average)}
\usage{
simulateProxPathGroup(model, pathway)
}
\arguments{
\item{model}{A CellModel}

\item{pathway}{A gene pathway}
}
\value{
the size of the cell population over time
}
\description{
\code{simulateProxPathGroup} Simulate Proximity Based Gene Expression (Average)
}

